[feat] Enum 정의
